DDIService/requirements.txt
# Core ML and scientific computing
torch>=2.0.0,<2.3.0
numpy>=1.21.0,<1.26.0
pandas>=1.3.0,<2.1.0
scikit-learn>=1.0.0,<1.4.0

# Chemistry and drug data
rdkit>=2022.9.0
PyTDC>=0.4.0

# Web framework
fastapi>=0.104.0,<0.110.0
uvicorn[standard]>=0.24.0,<0.25.0

# Validation and utilities
pydantic>=2.0.0,<3.0.0
tqdm>=4.64.0
requests>=2.31.0

# Configuration management
PyYAML>=6.0.0

# Visualization and evaluation
matplotlib>=3.5.0,<3.9.0
seaborn>=0.11.0,<0.13.0

# Development and testing (optional)
pytest>=7.0.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0

# Additional utilities
python-multipart>=0.0.6  # For file uploads in FastAPI
aiofiles>=23.0.0  # For async file operations
aiohttp>=3.8.0  # For async HTTP requests